Dynamically build attribute list
Now, instead of searching for a handful of hard-coded attributes and
including them in the <outline>, gather all property keys for a given
headline and include them -- and their values -- as attributes.

Only caveat is the property key can't be all upcase. Org includes a
handful of special properties as all upcase and we filter them away.