Accept OPML 1.0 and some other things
Not sure how safe OPML 1.0 is with conversion, but it seems to work with
output from at least one tool that exports it.

This also accepts an encoding declaration on the xml header, and
steps past a comment intervening between the header and the start
of OPML (I'm not familiar with Emacs regexp, the latter tweak may not
really be needed).

Signed-off-by: Eric Davis <eric@davising.com>
Closes #5