Preserve Org markup
This preserves Org-mode inline markup for italics, bold, entities, and
a few other things. Preserving markup avoids the need for
re-conversion when round-tripping of content from Org to OPML and
back, but it might not be the best choice for all tools with which
the OPML is shared. I guess eventually it would be good to have a
set of menu options for the treatment of markup on export.

Closes #7